create a debounce function
Explanation of Each Part:

The debounce function takes two arguments:
func: The function you want to debounce (e.g., console.log in your example).
delay: The amount of time (in milliseconds) to wait after the last invocation before calling func.

The Timeout Variable
let timeout;:
This variable will store the ID of the current timeout, which is used by clearTimeout() to cancel any previously scheduled executions of func.

The Inner Returned Function
The debounce function returns a new function (the debounced function). This returned function takes ...args (any arguments passed to it) and processes them.

...args: This syntax means "take all arguments passed to the function" and store them in the args array.

Clearing the Timeout
clearTimeout(timeout);:
Each time the debounced function is called, the clearTimeout(timeout) line clears the previous timeout (if any). This ensures that the func will not be executed before the full delay time, effectively resetting the timer each time the function is called.

Setting a New Timeout
timeout = setTimeout(() => { func.apply(this, args); }, delay);:

After clearing any previous timeout, we set a new one.
setTimeout() will execute the function after the delay time has passed.

func.apply(this, args) calls the original function (func) with the context (this) and the arguments (args) from the debounced function.

The reason apply(this, args) is used is to ensure that the function is called with the correct context and arguments.